Move EventThread to infrastructure:async module and convert it to use an AsyncRunner.
