+++ sync/src/test/java/tech/pegasys/teku/sync/multipeer/PeerChainTrackerTest.java
@@ -25,6 +25,8 @@ import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 import tech.pegasys.teku.datastructures.blocks.SlotAndBlockRoot;
 import tech.pegasys.teku.datastructures.util.DataStructureUtil;
+import tech.pegasys.teku.infrastructure.async.eventthread.EventThread;
+import tech.pegasys.teku.infrastructure.async.eventthread.InlineEventThread;
 import tech.pegasys.teku.networking.eth2.peers.Eth2Peer;
 import tech.pegasys.teku.networking.eth2.peers.Eth2Peer.PeerStatusSubscriber;
 import tech.pegasys.teku.networking.eth2.peers.PeerStatus;
@@ -32,8 +34,6 @@ import tech.pegasys.teku.networking.p2p.network.P2PNetwork;
 import tech.pegasys.teku.networking.p2p.peer.PeerConnectedSubscriber;
 import tech.pegasys.teku.networking.p2p.peer.PeerDisconnectedSubscriber;
 import tech.pegasys.teku.sync.multipeer.chains.TargetChain;
-import tech.pegasys.teku.sync.multipeer.eventthread.EventThread;
-import tech.pegasys.teku.sync.multipeer.eventthread.InlineEventThread;
 
 class PeerChainTrackerTest {
   private final DataStructureUtil dataStructureUtil = new DataStructureUtil();
